Use mini-language for cut ami
Implements a mini-language for "cut_from_edp", aka the "cut ami"
command that's a little more user-friendly than the previous regex-based
parsing method. Specifically, the "using" and "with" statements can now
appear in any order, as can the "verbose" and "noop" options. Also, the
mini-language uses familiar whitespace-agnostic semantics for token
separation, so things like "with configuration = master" are now valid.
Best of all, Alton will now report an error for "cut ami" commands with
invalid syntax that includes the position of the error and the
expected token.

Also removes the function supporting the old version of the "cut ami"
command.